+++ Pulsarc/Utils/Audio/AudioManager.cs
@@ -31,8 +31,7 @@ namespace Pulsarc.Utils.Audio
         {
             Stop();
 
-            if (songPath == "")
-                return;
+            if (songPath == "") { return; }
 
             // Initialize the song
             try
@@ -45,7 +44,7 @@ namespace Pulsarc.Utils.Audio
             }
             catch (AudioEngineException)
             {
-                PulsarcLogger.Debug(ManagedBass.Bass.LastError.ToString(), LogType.Runtime);
+                PulsarcLogger.Error(ManagedBass.Bass.LastError.ToString(), LogType.Runtime);
             }
 
             song.ApplyRate(Config.GetBool("Audio", "RatePitch"));
@@ -73,14 +72,13 @@ namespace Pulsarc.Utils.Audio
             threadLimiterWatch.Start();
             activeThreadLimiterWatch = true;
 
-            if (songPath == "")
-                return;
+            if (songPath == "") { return; }
 
             running = true;
             var threadTime = new Stopwatch();
 
             // Initialize the song
-            song = new AudioTrack(songPath, false)
+            song = new AudioTrack(songPath)
             {
                 Rate = audioRate,
                 Volume = Config.GetInt("Audio", "MusicVolume"),
@@ -183,9 +181,13 @@ namespace Pulsarc.Utils.Audio
                 Pause();
 
                 if (song.IsPlaying)
+                {
                     song.Stop();
+                }
 
+                song.Dispose();
                 song = null;
+                
                 Reset();
             }
         }